% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsm_mask.R
\name{dsm_mask}
\alias{dsm_mask}
\title{Mask disimilarity matrix with NAs for the off-diagonal elements}
\usage{
dsm_mask(dsm, lower = FALSE, melt = TRUE)
}
\arguments{
\item{dsm}{a single matrix}

\item{lower}{logical indicating selection of lower or upper diagonal}

\item{melt}{logical indicating whether or not to melt the matrix into a data frame}
}
\description{
This function will subsitute NAs in the off diagonal elements of a dissimilarity matrix. Will output melted matrix for convienent ggploting, or can simply mask an existing matrix.
}
\examples{
cars_dsm <- 1-cor(mtcars)
dsm_mask(cars_dsm)
}
\author{
Robert S. Chavez
}
