% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsm_vec.R
\name{dsm_vec}
\alias{dsm_vec}
\title{Vectorize Dissimilarity Matrix}
\usage{
dsm_vec(dsm, lower = TRUE)
}
\arguments{
\item{dsm}{a single matrix}

\item{lower}{logical indicating selection of lower or upper diagonal}
}
\description{
Convert dissimilarity matrix into off-diagonal vector
}
\examples{
cars_dsm <- 1-cor(mtcars)
dsm_vec(cars_dsm)
}
\author{
Robert S. Chavez
}
