% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbarmaker.R
\name{ggbarmaker}
\alias{ggbarmaker}
\title{Easy Bar Graph Maker for ggplot2}
\usage{
ggbarmaker(dv, groups, factor = NULL, id = NULL,
  groups_between = FALSE, palette = "Set1")
}
\arguments{
\item{dv}{Dependent variable vector}

\item{groups}{Groups or first factor variable.}

\item{factor}{Second factor variable.}

\item{id}{The name of a varible that identifieseach subject for withins-subject designs.}

\item{groups_between}{The 'groups' variable is a between subjects factor.}

\item{palette}{Select RColorBrewer palette.}
}
\description{
Makes quick and easy bar graphs for ggplot2 without pre-summarizing.
}
\note{
You must manually adjust y-axis range using the ggplot2 function coord_cartesian().
}
\examples{
# Between subject design.
x <- rnorm(100,100,15)
y <- sample(c("Group 1","Group 2"),100,TRUE)
z <- sample(c("Low","Med","High"),100,TRUE)
# Only one factor.
ggbarmaker(x,y)

# Two factors.
ggbar.maker(x,y,z)



# Within subject design
x <- rnorm(100,100,15)
y <- sample(c("Group 1","Group 2"),100,TRUE)
z <- sample(c("Low","Med","High"),100,TRUE)
subject <- rep(1:5,20)

# No between subjects factors.
ggbarmaker(x,y,z, id =subject)

# Grouping factor is between subjects.
ggbarmaker(x,y,z, id =subject, groups_between=TRUE)
}
\author{
Robert S. Chavez
}
