% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Multiple ggplots in a single plot}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{ggplot objects}

\item{plotlist}{Use a list of ggplot objects.}

\item{file}{Use ggplots from file.}

\item{cols}{Number of columns in multiplot.}

\item{layout}{Order layout.}
}
\description{
Print multiple graphs from ggplot2 on one page.
}
\note{
See : http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/
}
\examples{
x <- rnorm(100,100,15)
y <- sample(c("Group 1","Group 2"),100,TRUE)
p1 <- qplot(x)
p2 <- qplot(y)
multiplot(p1,p2)
}
\author{
Winston Chang
}
