% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normDataWithin.R
\name{normDataWithin}
\alias{normDataWithin}
\title{Normalizing helper function for within-subjects data..}
\usage{
normDataWithin(
  data = NULL,
  idvar,
  measurevar,
  betweenvars = NULL,
  na.rm = FALSE,
  .drop = TRUE
)
}
\arguments{
\item{data}{A data frame.}

\item{idvar}{The name of a column that identifies each subject (or matched subjects).}

\item{measurevar}{The name of a column that contains the variable to be summariezed.}

\item{betweenvars}{A vector containing names of columns that are between-subjects variables.}

\item{na.rm}{A boolean that indicates whether to ignore NA's.}
}
\description{
Norms the data within specified groups in a data frame; it normalizes each.
}
\note{
See : http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/
}
\examples{
x <- rnorm(100,100,15)
y <- sample(c("Group 1","Group 2"),100,TRUE)
z <- sample(c("Low","Med","High"),100,TRUE)
subject <- rep(1:5,20)
df <- data.frame(x,y,z,subject)

summarySEwithin(data=df, measurevar="x", betweenvars="y", withinvars="z", idvar="subject")
}
\author{
Winston Chang
}
