% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_factor_levels.R
\name{rev_factor_levels}
\alias{rev_factor_levels}
\title{Reverse Factor Levels}
\usage{
rev_factor_levels(x)
}
\arguments{
\item{x}{A single factor vector.}
}
\description{
This function reverses factor levels for using in plotting.
}
\examples{
rev_factor_levels(iris$Species)
}
\keyword{factor}
\keyword{levels}
\keyword{reverse}
