% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_score.R
\name{reverse_score}
\alias{reverse_score}
\title{Reverse Score Survey Item}
\usage{
reverse_score(x, xmax, negvals = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{xmax}{maximum value of scale item}

\item{negvals}{logical indicating if the scale contain negative values}
}
\description{
Calculates the reverse scoring of a standard survey item. You must specify the maximum value of the survey item if the values are all postive (e.g. 7 on a 1 to 7 scale). You may also specify if a scale's range include negative values that straddle zero (e.g. -5 to 5 scales).
}
\examples{
x <- sample(1:7, 10, replace = TRUE)
reverse_score(x, 7)

y <- sample(-5:5, 10, replace = TRUE)
reverse_score(y, negvals = TRUE)
}
\author{
Robert S. Chavez
}
