% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_agg.R
\name{rr_agg}
\alias{rr_agg}
\title{Round-robin single subject aggregate correlation distance calculator}
\usage{
rr_agg(self, tar, groups, cor_meth = "spearman")
}
\arguments{
\item{self}{The 'subject' column from the rr_selfonly data frame from rr_df_maker.}

\item{tar}{The 'target' column from the rr_noself_noNA data frame from rr_df_maker.}

\item{cor_meth}{Correlation method from cor() used in calculation. Default = 'spearman'.}

\item{group}{The 'group' column from the rr_noself_noNA data frame from rr_df_maker.}
}
\description{
Calculates the correlation distance between subject variables and target variables
after aggregated across raters. Self-to-self ratings are not included in the aggregated calculations.
This function works on a single subject basis, which is sometimes useful. It is also called
in rr_agg_all() for use over the whole data frame.
}
\note{
This function relies on the output from rr_df_maker().
}
\examples{
rr_agg(1,1,1) # self-congruent
rr_agg(1,2,1) # self-incongruent
}
\author{
Robert S. Chavez
}
