% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_agg_all.R
\name{rr_agg_all}
\alias{rr_agg_all}
\title{Round-robin whole data frame aggregate correlation distance analysis}
\usage{
rr_agg_all(subs_df, cor_meth = "spearman")
}
\arguments{
\item{subs_df}{The 'subs' data frame from rr_df_maker() output.}

\item{cor_meth}{Correlation method from cor() used in calculation. Default = 'spearman'.}
}
\description{
Calculates the correlation distance between subject variables and target variables
after aggregated across raters for an entire subject list in a data frame. Outputs
a data frame with aggregated correlation information and congruent/incongruent information.
Self-to-self ratings are not included in the aggregated calculations. This function relies
on the output from rr_df_maker().
}
\note{
This function relies on the output from rr_df_maker().
}
\examples{
rr_agg_all(subs, cor_meth = spearman)
}
\author{
Robert S. Chavez
}
