% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_agg_plot.R
\name{rr_agg_plot}
\alias{rr_agg_plot}
\title{Round-robin aggregated plot maker}
\usage{
rr_agg_plot(consensus_df)
}
\arguments{
\item{consensus_df}{data frame in the form from rr_agg_all().}
}
\description{
Creates plot for aggregated round-robin recapitulation analysis
}
\note{
Requires output from rr_agg_all()
}
\examples{
rr_agg_plot(consensus_df)
}
\author{
Robert S. Chavez
}
