% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_df_maker.R
\name{rr_df_maker}
\alias{rr_df_maker}
\title{Round-robin data frame maker}
\usage{
rr_df_maker(fsl_zstats, subs_csv)
}
\arguments{
\item{fsl_zstats}{A path to a text file from fslmeants output.}

\item{subs_csv}{A .csv file with group, subject, and target columns that will be used in the analysis.}
}
\description{
Creates a set of data frames for use in a CSNL round-robin analyses of fMRI data.
Output from this function will be necessary for out rr_* functions in this package.
}
\examples{
rr_df_maker("~/data/fsl_output.txt", "~/data/subect_list.csv")
}
\author{
Robert S. Chavez
}
