% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_error.R
\name{std_error}
\alias{std_error}
\title{Standard Error Calculator}
\usage{
std_error(x, na.rm = FALSE)
}
\arguments{
\item{x}{Numeric vector.}

\item{na.rm}{logical indicating removal of NAs before calculation.}
}
\description{
Calculates standard error of the mean for numeric vector.
}
\examples{
x <- rnorm(100,100,15)
std_error(x)
}
\author{
Robert S. Chavez
}
