% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarySE.R
\name{summarySE}
\alias{summarySE}
\title{Summarize data for between-subjects designs.}
\usage{
summarySE(data = NULL, measurevar, groupvars = NULL, na.rm = FALSE,
  conf.interval = 0.95, .drop = TRUE)
}
\arguments{
\item{data}{A data frame.}

\item{measurevar}{The name of a column that contains the variable to be summariezed.}

\item{na.rm}{A boolean that indicates whether to ignore NA's.}

\item{conf.interval}{The percent range of the confidence interval, 95 percent default.}

\item{betweenvars}{A vector containing names of columns that are between-subjects variables.}

\item{idvar}{The name of a column that identifies each subject (or matched subjects).}
}
\description{
Gives count, mean, standard deviation, standard error of the mean, and confidence interval (default 95 percent)
}
\note{
See : http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/
}
\examples{
x <- rnorm(100,100,15)
y <- sample(c("Group 1","Group 2"),100,TRUE)
df <- data.frame(x,y)
summarySE(data=df, measurevar="x", betweenvars="y")
}
\author{
Winston Chang
}
