% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarySEwithin.R
\name{summarySEwithin}
\alias{summarySEwithin}
\title{Summarize data for within-subjects designs.}
\usage{
summarySEwithin(
  data = NULL,
  measurevar,
  betweenvars = NULL,
  withinvars = NULL,
  idvar = NULL,
  na.rm = FALSE,
  conf.interval = 0.95,
  .drop = TRUE
)
}
\arguments{
\item{data}{A data frame.}

\item{measurevar}{The name of a column that contains the variable to be summariezed.}

\item{betweenvars}{A vector containing names of columns that are between-subjects variables.}

\item{withinvars}{A vector containing names of columns that are within-subjects variables}

\item{idvar}{The name of a column that identifies each subject (or matched subjects).}

\item{na.rm}{A boolean that indicates whether to ignore NA's.}

\item{conf.interval}{The percent range of the confidence interval, 95 percent default.}
}
\description{
Summarizes data, handling within-subjects variables by removing inter-subject variability. It will still work if there are no within-S variables. Gives count, un-normed mean, normed mean (with same between-group mean), standard deviation, standard error of the mean, and confidence interval. If there are within-subject variables, calculate adjusted values using method from Morey (2008).
}
\note{
See : http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/
}
\examples{
x <- rnorm(100,100,15)
y <- sample(c("Group 1","Group 2"),100,TRUE)
z <- sample(c("Low","Med","High"),100,TRUE)
subject <- rep(1:5,20)
df <- data.frame(x,y,z,subject)
summarySEwithin(data=df, measurevar="x", betweenvars="y", withinvars="z", idvar="subject")
}
\author{
Winston Chang
}
