% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsm_plot.R
\name{dsm_plot}
\alias{dsm_plot}
\title{Off-diagonal Plotting Function for Dissimlarity Matricies}
\usage{
dsm_plot(df, col_n = 8, color_dir = 1, rev_fac = FALSE, melt = TRUE)
}
\arguments{
\item{melt}{logical indicating whether or not to melt the matrix into a data frame prior to plotting}

\item{dsm}{a single matrix}

\item{lower}{logical indicating selection of lower or upper diagonal}
}
\description{
This function will plot the off-diagonal elements of a dissmilarity matrix.
}
\examples{
cars_dsm <- 1-cor(mtcars)
dsm_plot(cars_dsm)
}
\author{
Robert S. Chavez
}
